+++ projects/_03_independent_cqc/utils/tests/test_utils.py
@@ -4,8 +4,12 @@ import warnings
 from pyspark.sql import Window
 
 import projects._03_independent_cqc.utils.utils.utils as job
-from tests.test_file_data import IndCQCDataUtils as Data
-from tests.test_file_schemas import IndCQCDataUtils as Schemas
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
+    IndCQCDataUtils as Data,
+)
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
+    IndCQCDataUtils as Schemas,
+)
 from utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
